requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alabaster==0.7.12
argh==0.26.2
asgiref==3.2.3
babel==2.6.0
bumpversion==0.5.3
celery
certifi==2019.3.9
chardet==3.0.4
click==7.0
commonmark==0.9.0
coreapi
coreschema
coverage==4.5.1
django-debug-toolbar==1.11
django-environ==0.4.5
django-model-utils==3.1.2
django==2.2.1
djangorestframework==3.10.3
djangorestframework-xml
django-filter
dnspython==1.16.0
docutils==0.14
eparsecis==3.1.0
epcpyyes==2.0.1
faker==0.9.1
flake8==3.5.0
future==0.17.1
gunicorn==20.0.4
haikunator==2.1.0
idna==2.8
imagesize==1.1.0
importlib-metadata==1.4.0
jinja2==2.10.1
lxml==4.3.3
markupsafe==1.1.1
mccabe==0.6.1
mixer==6.1.3
more-itertools==8.1.0
packaging==19.0
pathtools==0.1.2
pip-tools==4.4.0
pkginfo==1.5.0.1
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.1
pycodestyle==2.3.1
pyflakes==1.6.0
pygments==2.4.0
pyparsing==2.4.0
python-dateutil==2.8.0
python-dotenv==0.10.5
pytz==2019.1
pyyaml==5.3
quartet-capture==2.6.5
quartet-epcis==1.4.11
quartet-masterdata==1.3.0
recommonmark==0.5.0
requests-toolbelt==0.9.1
requests==2.22.0
rstcheck==3.3.1
six==1.12.0
snowballstemmer==1.2.1
sphinx-rtd-theme==0.4.3
sphinx==2.0.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
sqlparse==0.3.0
text-unidecode==1.2
tox==2.9.1
tqdm==4.41.1
twine==1.10.0
urllib3==1.25.3
virtualenv==16.7.9
watchdog==0.8.3
zipp==2.0.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools
